requirements.txt
python-dotenv==1.0.0
prometheus_client==0.16.0
requests==2.25.1